feat(gonx)!: remove main.go as the target of the build and run commands (#79)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [x] The commit message follows our guidelines: CONTRIBUTING.md#commit
- [x] Tests for the changes have been added (for bug fixes / features)
- [x] Docs have been added / updated (for bug fixes / features)

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

```
[ ] Bugfix
[x] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Build related changes
[ ] CI related changes
[ ] Documentation content changes
[ ] Other... Please describe:
```

## What is the current behavior?

We run the `go build` and `go run` commands over the `main.go` file of
the project.

## What is the new behavior?

We run the `go build` and `go run` commands over the directory itself
using '.' relative path. This is the go way.

## Does this PR introduce a breaking change?

```
[x] Yes
[ ] No
```

The `main` file configuration is no longer accepted.