feat: add generate executor (#83)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [x] The commit message follows our guidelines: CONTRIBUTING.md#commit
- [x] Tests for the changes have been added (for bug fixes / features)
- [x] Docs have been added / updated (for bug fixes / features)

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

```
[ ] Bugfix
[x] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Build related changes
[ ] CI related changes
[ ] Documentation content changes
[ ] Other... Please describe:
```

## What is the current behavior?

We had removed the `generate` executor after forking the nx-go repo,
thinking it was a way to generate code similar to how nx does it.

Fixes: #82 

## What is the new behavior?

We have implemented the `generate` executor and made it a dependency for
both he test and build executors.

## Does this PR introduce a breaking change?

```
[x] Yes
[ ] No
```

Now the build and test executors could fail if the generate executor
fails.