docs: fix gonx-docs nx configuration (#88)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [x] The commit message follows our guidelines: CONTRIBUTING.md#commit
- [x] Tests for the changes have been added (for bug fixes / features)
- [x] Docs have been added / updated (for bug fixes / features)

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

```
[ ] Bugfix
[ ] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Build related changes
[ ] CI related changes
[x] Documentation content changes
[ ] Other... Please describe:
```

## What is the current behavior?

gonx-docs was using the package.json to define the nx project
configuration but nx was not detecting it

## What is the new behavior?

We have replace the `package.json` with a `project.json` and now nx is
properly detecting the project.

We have also update to the latest minor version of Nx

## Does this PR introduce a breaking change?

```
[ ] Yes
[x] No
```